Move all headers to include, use .hpp for all headers
