+++ arm-controller/CMakeLists.txt
@@ -20,6 +20,11 @@ include_directories(include)
 
 scrb_add_executable(${PROJECT_NAME}_node src/arm_controller_node.cpp)
 
+target_include_directories(${PROJECT_NAME}_node PUBLIC
+        $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>
+        $<INSTALL_INTERFACE:include>
+)
+
 target_link_libraries(${PROJECT_NAME}_node PUBLIC ${DEPENDENCY_TARGETS})
 
 target_precompile_headers(${PROJECT_NAME}_node PRIVATE
@@ -28,6 +33,11 @@ target_precompile_headers(${PROJECT_NAME}_node PRIVATE
 )
 
 ## INSTALL
+
+install(
+        DIRECTORY include
+        DESTINATION include/${PROJECT_NAME}
+)
 
 install(
         TARGETS ${PROJECT_NAME}_node
