+++ arm-controller/src/arm_controller_node.cpp
@@ -1,4 +1,4 @@
-#include "arm_controller_node.h"
+#include "arm_controller/arm_controller_node.hpp"
 
 #include <cstring>
 #include <fcntl.h>
