+++ transducer-imu/include/EasyProfile/EasyQueue.hpp
