+++ transducer-imu/src/EasyProfile/EasyProtocol.cpp
@@ -3,7 +3,7 @@
  * @author COPYRIGHT(c) 2017 SYD Dynamics ApS
  * @see    EasyProtocol.h for more comments
  */
-#include "EasyProfile/EasyProtocol.h"
+#include "EasyProfile/EasyProtocol.hpp"
 
 
 EasyProtocol::EasyProtocol() {
