+++ transducer-imu/src/serialib/serialib.cpp
@@ -15,7 +15,7 @@ FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN 
 This is a licence-free software, it can be used by anyone who try to build a better world.
  */
 
-#include "serialib/serialib.h"
+#include "serialib/serialib.hpp"
 
 
 //_____________________________________
