+++ wheels-controller/src/test.cpp
@@ -1,7 +1,7 @@
 //
 // Created by nik on 08/02/24.
 //
-#include "command_prefixes.h"
+#include "wheels_controller/command_prefixes.hpp"
 #include <cstdint>
 #include <cstdio>
 #include <cstring>
@@ -169,7 +169,7 @@ int main() {
                 std::cerr << errno;
                 exit(-1);
             }
-            /////////////////////////////////////////////////////////////////////////////////////////////////////////////        
+            /////////////////////////////////////////////////////////////////////////////////////////////////////////////
             frame.can_id = (COMMAND_PREFIX_VELOCITY_CONTROL << 8) | (deviceID_2 + 0x80);
             frame.can_dlc = 8;
             frame.can_id |= CAN_EFF_FLAG;
